+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.2.11 (unreleased)
+## 0.2.11
 
 - Fix idempotency issue with various resource and boolean values
 - Remove redundant `per_tunnel_qos` and `per_tunnel_qos_aggregator` attributes from `sdwan_cisco_vpn_interface_feature_template` resource and data source